+++ libs/Core/Foundations.php
@@ -4,7 +4,7 @@ namespace libs\Core;
 
 class Foundations
 {
-    const VERSION = '1.0.35';
+    const VERSION = '1.0.36';
 
     /**
      * Get the version number of the application.